unify `for_each_relevant_impl`
