+++ tests/ui/methods/inherent-bound-in-probe.stderr
@@ -12,14 +12,14 @@ LL | struct Helper<'a, T>
 note: required by a bound in `std::iter::IntoIterator::IntoIter`
   --> $SRC_DIR/core/src/iter/traits/collect.rs:LL:COL
 
-error[E0275]: overflow evaluating the requirement `&_: IntoIterator`
+error[E0275]: overflow evaluating the requirement `&HashMap<_, _, _, _>: IntoIterator`
   --> $DIR/inherent-bound-in-probe.rs:43:9
    |
 LL |         Helper::new(&self.0)
    |         ^^^^^^
    |
    = help: consider increasing the recursion limit by adding a `#![recursion_limit = "256"]` attribute to your crate (`inherent_bound_in_probe`)
-note: required for `&BitReaderWrapper<_>` to implement `IntoIterator`
+note: required for `&BitReaderWrapper<HashMap<_, _, _, _>>` to implement `IntoIterator`
   --> $DIR/inherent-bound-in-probe.rs:33:13
    |
 LL | impl<'a, T> IntoIterator for &'a BitReaderWrapper<T>
